pyproject.toml
[tool.poetry]
name = "jamesneedsdata"
version = "0.1.0"
description = ""
authors = ["kenny <eldridge.kenneth@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.3"
click = "^8.1.7"
seaborn = "^0.13.0"
poetry = "^1.7.1"
urllib3 = "^2.1.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
